requirements.txt
Django==4.1.3
Pillow==9.3.0
python-decouple==3.6
django-crispy-forms==1.14.0
djangorestframework~=3.14.0